split out class into its own file
